feat: String 클래스에 대한 학습 테스트 - 요구사항 3 시도 중
